+++ PKHeX.Core/PKM/Shared/PokeList1.cs
@@ -2,7 +2,7 @@ namespace PKHeX.Core
 {
     public sealed class PokeList1 : PokeListGB<PK1>
     {
-        protected override byte GetSpeciesBoxIdentifier(PK1 pk) => (byte)SpeciesConverter.SetG1Species(pk.Species);
+        protected override byte GetSpeciesBoxIdentifier(PK1 pk) => SpeciesConverter.SetG1Species(pk.Species);
         protected override PK1 GetEntry(byte[] dat, byte[] otname, byte[] nick, bool egg) => new PK1(dat) { otname = otname, nick = nick };
         protected override int GetEntrySize() => GetEntrySize(IsFormatParty);
